Improve lecture version update dialog
Improve lecture version update dialog
