+++ koin/src/main/java/in/koreatech/koin/ui/timetable/presenter/TimetableAnonymousContract.java
@@ -45,7 +45,7 @@ public interface TimetableAnonymousContract{
 
         void showFailSavedTimeTable();
 
-        void showUpdateAlertDialog(String message);
+        void showUpdateAlertDialog(String serverVersionCode);
 
         void updateSemesterCode(String semester);
